+++ frontend/src/components/AdminAppbar.jsx
@@ -1,12 +1,14 @@
-import React from 'react'
+import React, { useState } from 'react'
 import SupervisorAccountIcon from '@mui/icons-material/SupervisorAccount';
 import InventoryIcon from '@mui/icons-material/Inventory';
 import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
 import { Box, Drawer, AppBar, Toolbar, List, Typography, ListItem, ListItemIcon, ListItemText } from '@mui/material';
+import AdminCustomerList from './AdminCustomerList';
 const drawerWidth = 210;
 
 
-export default function AdminAppbar() {
+export default function AdminAppbar({handleItemClick}) {
+
   return (
     <div>
         <AppBar position="fixed" sx={{ zIndex: (theme) => theme.zIndex.drawer + 1 }}>
@@ -28,8 +30,8 @@ export default function AdminAppbar() {
                 <Toolbar />
                 <Box sx={{ overflow: 'auto' }}>
                     <List>
-                    {[['Products', <InventoryIcon/>], ['Order',<ShoppingCartIcon/>], ['Customers', <SupervisorAccountIcon/>]].map((item, index) => (
-                        <ListItem button key={item[0]}>
+                    {[['Products', <InventoryIcon/>], ['Orders',<ShoppingCartIcon/>], ['Customers', <SupervisorAccountIcon/>]].map((item, index) => (
+                        <ListItem button key={item[0]} onClick={() => handleItemClick(item[0])}>
                         <ListItemIcon>
                             {item[1]}
